reorganized some files, removed some unused code, prompting for Google acct on every login
