+++ client/js/containers/TrailDetail.js
