+++ client/js/containers/index.js
@@ -1,4 +1,4 @@
-export Main from './Main'
+export TrailDetail from './TrailDetail'
 export TrailFavorites from './TrailFavorites'
 export TrailList from './TrailList'
 export TrailsMain from './TrailsMain'
