+++ server/routes/auth.js
@@ -7,7 +7,10 @@ const router = express.Router()
 router.get('/google',
   passport.authenticate(
     'google',
-    {scope: ['profile']}
+    {
+      scope: ['profile'],
+      prompt: 'select_account'
+    },
   )
 )
